+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased](https://github.com/sdss/sdsstools/compare/0.1.8...HEAD)
 
+### Fixed
+
+- Handle case where log header is None.
+
 ## [0.1.8](https://github.com/sdss/sdsstools/compare/0.1.7...0.1.8)
 
 ### Added